Add a form and input with specific attributes
I added a web form to collect information from users.
The action attribute indicates where form data or path should be sent.

The input element allows the developer several ways to collect data from a web form.

There are many kinds of inputs you can create using the type attribute. Namely, a password field, reset button, or a control to let users select a file from their computer.

For a form's data to be accessed by the location specified in the action attribute, give the text field a name attribute and assign it a value to represent the data being submitted. 

Placeholder text is used to give people a hint about what kind of information to enter into an input.

Add the required attribute to an input element to prevent a user from submitting the form when required information is missing.