Add radio buttons with necessary attributes
Use radio buttons for questions where you want only one answer out of multiple options.

<label> elements are used to help associate the text for an input element with the input element itself (especially for assistive technologies like screen readers) makes it so clicking the word cat also selects the corresponding radio button.

The id attribute is used to identify specific HTML elements. Each id attribute's value must be unique from all other id values for the entire page.

To make it so that selecting one radio button automatically deselects the other, both buttons must have a name attribute with the same value.

If the user selects the Indoor radio button and submit the form, the form data for the button is based on its name and value attributes. 
If the radio button doesn't have a value attribute, the form data will include indoor-outdoor=on, which is not useful when you have multiple buttons.
I used the same value of the id attribute as the value for the value attribute.