Create another fieldset with checkboxes
Forms commonly use checkboxes for questions that may have more than one answer.

Another way to associate an input element's text with the element itself is to nest the text within a label element and add a for attribute with the same value as the input element's id attribute.

Add a name attribute with the value personality. This makes it easier for a server to process the web form, especially when there are multiple checkboxes.

Like radio buttons, form data for selected checkboxes are name / value attribute pairs. While the value attribute is optional, it's best practice to include it with any checkboxes or radio buttons on the page.

Make a checkbox checked or radio button selected by default, add the checked attribute to it. There's no need to set a value to the checked attribute.